moving data, adding README, turn on travis.
The data() command did not deal well with a raw csv. I have
moved to the use_data_raw() setup suggested by devtools,
where the raw data are in data-raw/svs.csv, and rda file is
in data. The data now load.

Which is visible in the shiny new README.

Not that it is needed, but try travis here.